backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
ffmpeg-python==0.2.0
python-dotenv==1.0.0
pydantic==2.4.2
numpy==1.26.4
torch==2.2.0
transformers==4.37.2
srt==3.5.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
googletrans==3.1.0a0
pysrt==1.1.2
webvtt-py==0.4.6
ass==0.5.2 